Created main and comms_n64_console modules
main.h and main.c have been created/changed to be more abstract and better organized. Specifically there is an init function and a function to control the state of the blue led. both files are tested and work.

comms_n64_console.h and comms_n64_console.c has been created to abstract away the uart module required to communicate with the n64 controller. nothing is tested yet in this module.