n64 stop bit function has been corrected
the implemented stop bit is now about 1.33us long in duration. when the code returns from sending a command to the controller, there is about 2.4us left to prepare for the returned information. the transmission is also very very steady as the controller seems to respond to a request every time. it looks like the stop bit was the culprit for dropped packets before. next feature will likely be to parse the information returned from the controller.