uart rx problem
it looks like after receiving 12 uart bytes (3 n64 bytes), the code gets stuck in a loop and thinks there is no data in the rx buffer and loops forever. it actually lines up where you would need to read the y axis value. i think this hints at a tolerance issue with the baudrate timing. the baudrate might need to be nudged a little more (up or down?) to land in a favorable spot tolerance wise. this could also explain the last noted issue of the data received differs sometimes.