uart rx problemed half resolved
changing the baudrate from 1250000 to 1300000 resolved the timing issue where the uart buffer would get stuck and loop forever. the issue was resolved by first going below 1250000 in arbitrary increments until hitting 1000000. when nothing was working or working worse, it was then incremented to 1300000. this value is likely not optimized. it definitely fixes the "getting stuck issue and it does improve the integrity of the bytes received. however, there are some instances where the uart data received is outside of the expected set of  0x08, 0xE8, 0x0F, and 0xEF. for example, some times 0xC8 is received instead of 0xE8. this is also baudrate related.

though its safe to say the code will never get stuck. it was ran for over 20 minutes without a problem. a great improvement over the last record of 9 seconds. a bit more tweaking on the baudrate and the uart rx problems will completely be solved.