uart rx filter + parser
tried to optimize the baud rate so there are no timing errors but it was not possible to easily gets this value. however, there are only two value possible now per bit pair and they are mutually exclusive of each other. that means even with a bit timing, it doesn't matter as it won't interfere with other bit pairs. this logic is reflected in the code.

a parsing function has been created. it parses the controller response bytes and neatly packs them into an array that better illustrates on the n64 controller contains its data. the parsing algorithm is not pleasant to look at and needs to be better coded later on. but the goal was to make sure everything was parse-able and that the alues could be read back. it looks like the buttons work fine but i need to double check the x and y axis values to make sure the order are what is expected.