Abstracted some constants & created BBG compatible registers
Some of the byte, register, and bit locations were hardcoded in, so decided to place them in a macro to make the code more readable.

Created a function to properly set the data from the n64 controller parsed data. This means that these new registers can now be passed easily over uart to a BBG control board that has an expansion port. These registers were echoed out on a uart and observed on a logic analyzer and have been verified to be working okay. So far only the digital buttons have been worked on. Saving the x and y axis of the control stick to be the last thing to implement.

Next thing to do would be setting up a separate uart to pass this data to a BBG control board. The uart mentioned before is still the one meant only for n64 controller communication.